adaptation of IgDetective to all loci
